A 1st attempt to use the mat-grid-list as the basis of a holy-grail layout implementation
I created a new branch -the ‘matgridlist’ branch for the commit that only pure Angular Material has had just installed.

The most important challenge was the alignment of items inside a mat-grid-tile.
A working solution was to use a <span> element inside the mat-grid-tile, and assign to that span the specific class. Then in the specific class I used a flex container. Note, that font-size and color (background-color) was working OK, but to achieve a desirable alignment of the component enclosed, I had to set height and width to 100%.